requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pyftdi==0.52.9
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.1
pyusb==1.1.1
six==1.15.0
